+++ src/EFCore.Ydb/src/Storage/Internal/YdbExecutionStrategy.cs
@@ -7,7 +7,7 @@ using Microsoft.EntityFrameworkCore.Storage;
 using Ydb.Sdk;
 using Ydb.Sdk.Ado;
 using Ydb.Sdk.Ado.RetryPolicy;
-using Ydb.Sdk.Tracing;
+using Ydb.Sdk.Ado.Tracing;
 
 namespace EntityFrameworkCore.Ydb.Storage.Internal;