+++ src/Ydb.Sdk/src/Ado/Session/RetryableSession.cs
@@ -2,7 +2,7 @@ using System.Diagnostics;
 using Ydb.Query;
 using Ydb.Sdk.Ado.Internal;
 using Ydb.Sdk.Ado.RetryPolicy;
-using Ydb.Sdk.Tracing;
+using Ydb.Sdk.Ado.Tracing;
 
 namespace Ydb.Sdk.Ado.Session;